Fix #7 Disabled status change when editing final invoice
Changing status on final invoice ended with an exception, so I have
disabled it. Updated logic to change invoice state in actions.

This way user does not have ability to generate the exception and
it is possible to change invoice state and as benefit we can change
several at onece.